backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-genai>=0.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.26.0
pymupdf>=1.23.0
pytesseract>=0.3.10
pillow>=10.0.0
pandas>=2.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
httpx>=0.26.0  # For TestClient
